backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
PyJWT>=2.8
google-auth>=2.25
google-auth-oauthlib>=1.2
psycopg2-binary>=2.9
python-dotenv>=1.0
supabase>=2.0.0
boto3>=1.34.0
